// src/types/container.ts
import { ContentViewColumn } from "./contentViewColumn";

export interface Container {
    columns: ContentViewColumn[];
    contentViewID: number | null;
    contentDefinitionID: number | null;
    contentDefinitionName: string | null;
    referenceName: string | null;
    contentViewName: string | null;
    contentDefinitionType: number;
    requiresApproval: boolean | null;
    lastModifiedDate: string | null;
    lastModifiedOn: string | null;
    lastModifiedBy: string | null;
    isShared: boolean | null;
    isDynamicPageList: boolean | null;
    disablePublishFromList: boolean | null;
    contentViewCategoryID: number | null;
    contentViewCategoryReferenceName: string | null;
    contentViewCategoryName: string | null;
    title: string | null;
    defaultPage: string | null;
    isPublished: boolean | null;
    schemaTitle: string | null;
    allowClientSideSave: boolean | null;
    defaultSortColumn: string | null;
    defaultSortDirection: string | null;
    usageCount: number | null;
    isDeleted: boolean | null;
    enableRSSOutput: boolean | null;
    enableAPIOutput: boolean | null;
    commentsRecordTypeName: string | null;
    numRowsInListing: number | null;
    contentDefinitionTypeID: number | null;
    fullSyncModDate: string | null;
    confirmSharingOnPublish: boolean | null;
    contentTemplateName: string | null;
    currentUserCanDelete: boolean | null;
    currentUserCanEdit: boolean | null;
    currentUserCanDesign: boolean | null;
    currentUserCanManage: boolean | null;
    currentUserCanContribute: boolean | null;
    currentUserCanPublish: boolean | null;
    defaultListingPage: string | null;
    defaultDetailsPage: string | null;
    defaultDetailsPageQueryString: string;
    isListItem: boolean | null;
} 